requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.0
pandas==2.1.3
numpy==1.25.2
matplotlib==3.8.1
plotly==5.17.0
flask==2.3.3
flask-cors==4.0.0
flask-socketio==5.3.6
python-socketio==5.9.0
schedule==1.2.0
pycryptodome==3.19.0
steampy==1.4.2
python-dotenv==1.0.0
sqlalchemy==2.0.23
aiohttp==3.9.1
asyncio==3.4.3
websockets==11.0.3
pillow==10.1.0